+++ CHANGELOG.md
@@ -15,6 +15,16 @@ but cannot always guarantee backwards compatibility. Changes that may **break co
   - 🟠 Renamed `RegressionModel` to `SKLearnModel`. Using `RegressionModel` will raise a depraction warning.
   - 🟠 Renamed `RandomForest` to `RandomForestModel`. Using `RandomForest` will raise a depraction warning.
   - 🔴 Renamed `RegressionModelWithCategoricalCovariates` to `SKLearnModelWithCategoricalCovariates`. Removed `RegressionModelWithCategoricalCovariates`
+- 🔴 Improvements to `TorchForecastingModel` datasets: [#2796](https://github.com/unit8co/darts/pull/2796) by [Dennis Bader](https://github.com/dennisbader).
+  - We simplified the training and inference datasets. Instead of having covariates specific datasets, the new datasets now support all combinations of covariates natively:
+    - `ShiftedTrainingDataset` (replaces all `*ShiftedDataset`)
+    - `SequentialTrainingDataset` (replaces all `*SequentialDataset`)
+    - `HorizonBasedTrainingDataset` (replaces `HorizonBasedDataset`)
+    - `SequentialInferenceDataset` (replaces all `*InferenceDataset`)
+  - All datasets now have uniform output:
+    - Training datasets: Tuple[past target, past cov, historic future cov, future cov, static cov, sample weight, future target].
+    - Inference datasets: Tuple[past target, past cov, future past cov, historic future cov, future cov, static cov, target TimeSeries, pred start time]
+  - `HorizonBasedTrainingDataset` now also supports future covariates.
 
 **Fixed**